+++ CHANGELOG.md
@@ -7,6 +7,8 @@ to keep track of.
 
 ## master
 
+## 0.1.4
+
 **Python**
 
 Features: